airflow/requirements.txt
requests
requests-oauthlib
apache-airflow==2.9.2 --constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.9.2/constraints-3.10.txt
vnstock3
ipython
boto3
pyspark==3.3.4
python-dotenv
papermill==2.3.4
apache-airflow-providers-apache-spark==4.10.0